added github workflow and test cases
