+++ .gitignore
@@ -42,6 +42,7 @@ nosetests.xml
 coverage.xml
 *.cover
 .hypothesis/
+.pytest_cache/
 
 # Translations
 *.mo
@@ -69,3 +70,5 @@ ENV/
 
 # mypy
 .mypy_cache/
+
+.vscode/
