+++ tests/test_json_utils.py
@@ -1,8 +1,6 @@
 import os
 import unittest
 import json
-
-from six import assertRaisesRegex
 from beacons_bio_3d.utils import JSONUtils, DjangoUtils, InvalidProviderException
 from jsonschema import ValidationError
 
@@ -25,11 +23,3 @@ class TestJSONUtils(unittest.TestCase):
     def test_invalid_providers_case(self):
         self.assertRaises(InvalidProviderException, JSONUtils.validate_schema, f"{RES_PATH}/valid_schema.json", f"{RES_PATH}/invalid_providers_registry.json")
 
-
-class TestDjangoUtils(unittest.TestCase):
-
-    def test_model_generate_valid_case(self):
-        registry_json = f"{RES_PATH}/valid_full_registry.json"
-        model_json = f"{RES_PATH}/valid_model.json"
-
-        generated_model = DjangoUtils.generate_fixture_json(registry_json)
